Move Sum, Product to own module
